requirements.txt
# Web Framework
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.1.3

# Async Task Queue
celery==5.4.0
redis==5.2.1

# Environment & Config
python-dotenv==1.0.1

# LLM Provider (OpenAI-compatible)
openai==1.75.0

# Speech Providers
deepgram-sdk==3.10.1
assemblyai==0.37.0
elevenlabs==2.3.0
httpx==0.28.1

# Audio Processing
pydub==0.25.1
ffmpeg-python==0.2.0

# Utilities
markdown==3.7
